refactor: 카테고리 API 호출방식 변경
