+++ src/api/resourses/categories.ts
@@ -1,0 +1,8 @@
+import axiosInstance from "../axiosInstance";
+
+// Categories API
+export const categoryAPI = {
+  getCategories: () => axiosInstance.get("/categories"),
+  getCategory: (categoryId: number) =>
+    axiosInstance.get(`/categories/${categoryId}`),
+};
